fix and update ray lightning tune example
