--- examples/ray-lightning-tune/README.md
@@ -1,4 +1,0 @@
-Get started with PyTorch Lightning and Ray on Metaflow:
-```
-python ray_lightning_tune_flow.py --environment=pypi run
-```
