+++ examples/ray_torch_lightning/dataloader.py
@@ -1,12 +1,13 @@
+import tempfile
+from filelock import FileLock
 import pytorch_lightning as pl
 from torchvision import transforms
 from torchvision.datasets import MNIST
 from torch.utils.data import DataLoader, random_split
-from filelock import FileLock
-import tempfile
 
 
 class MNISTDataModule(pl.LightningDataModule):
+
     def __init__(self, batch_size=128):
         super().__init__()
         self.data_dir = tempfile.mkdtemp()
@@ -15,22 +16,29 @@ class MNISTDataModule(pl.LightningDataModule):
             [transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))]
         )
 
-    def setup(self, stage=None):
+    def setup(self, stage: str):
         with FileLock(f"{self.data_dir}.lock"):
-            mnist = MNIST(
-                self.data_dir, train=True, download=True, transform=self.transform
-            )
+            mnist = MNIST(self.data_dir, train=True, download=True, transform=self.transform)
             self.mnist_train, self.mnist_val = random_split(mnist, [55000, 5000])
-
-            self.mnist_test = MNIST(
-                self.data_dir, train=False, download=True, transform=self.transform
-            )
+            self.mnist_test = MNIST(self.data_dir, train=False, download=True, transform=self.transform)
 
     def train_dataloader(self):
-        return DataLoader(self.mnist_train, batch_size=self.batch_size, num_workers=4)
+        return DataLoader(
+            self.mnist_train,
+            batch_size=self.batch_size,
+            num_workers=4
+        )
 
     def val_dataloader(self):
-        return DataLoader(self.mnist_val, batch_size=self.batch_size, num_workers=4)
+        return DataLoader(
+            self.mnist_val,
+            batch_size=self.batch_size,
+            num_workers=4
+        )
 
     def test_dataloader(self):
-        return DataLoader(self.mnist_test, batch_size=self.batch_size, num_workers=4)
+        return DataLoader(
+            self.mnist_test,
+            batch_size=self.batch_size,
+            num_workers=4
+        )
